Move localization files to the library folder
